+++ Makefile.am
@@ -89,6 +89,8 @@ ipfixprobe_process_src=\
 		process/sip.hpp \
 		process/tls.cpp \
 		process/tls.hpp \
+		process/tls_parser.cpp \
+		process/tls_parser.hpp \
 		process/smtp.cpp \
 		process/smtp.hpp \
 		process/dns-utils.hpp \
@@ -128,8 +130,9 @@ if WITH_QUIC
 ipfixprobe_process_src+=\
 		process/quic.hpp \
 		process/quic.cpp \
-		process/quic_implementation.cpp \
-		process/quic_implementation.hpp
+		process/quic_parser.cpp \
+		process/quic_parser.hpp 
+
 endif
 
 if WITH_FLEXPROBE
